# be/swagger.yaml
openapi: 3.0.3
info:
  title: Hexa Crochet E-commerce API
  description: |
    Backend API untuk toko e-commerce Hexa Crochet yang menjual produk rajutan (crochet) dengan menggunakan Express.js, Prisma ORM, dan Supabase PostgreSQL.
    
    ## Fitur Utama
    - Authentication dengan Supabase Auth
    - Manajemen produk dan kategori
    - Shopping cart dan checkout
    - Order management
    - Payment processing
    - Shipping tracking
    - Address management
    
    ## Authentication
    API menggunakan Supabase Auth dengan JWT tokens. Setiap request yang memerlukan autentikasi harus menyertakan header:
    ```
    Authorization: Bearer <access_token>
    ```
  version: 1.0.0
  contact:
    name: Hexa Crochet Team
    email: support@hexacrochet.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.hexacrochet.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/payments/paypal/create:
    post:
      tags:
        - Payments
      summary: Create PayPal payment
      description: Create a new PayPal payment for an order
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - order_id
              properties:
                order_id:
                  type: string
                  format: uuid
                  example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        '201':
          description: PayPal payment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: PayPal payment created successfully
                  data:
                    type: object
                    properties:
                      paypal_order_id:
                        type: string
                        example: 2X12345678901234567
                      approval_url:
                        type: string
                        example: https://www.sandbox.paypal.com/checkoutnow?token=2X12345678901234567
                      order_id:
                        type: string
                        format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/payments/paypal/capture:
    post:
      tags:
        - Payments
      summary: Capture PayPal payment
      description: Capture a PayPal payment after user approval
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - paypal_order_id
              properties:
                paypal_order_id:
                  type: string
                  example: 2X12345678901234567
      responses:
        '200':
          description: PayPal payment captured successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: PayPal payment captured successfully
                  data:
                    type: object
                    properties:
                      capture_id:
                        type: string
                        example: 1A12345678901234567
                      order_id:
                        type: string
                        format: uuid
                      status:
                        type: string
                        example: COMPLETED
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/payments/paypal/webhook:
    post:
      tags:
        - Payments
      summary: PayPal webhook handler
      description: Handle PayPal webhook events
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  example: 8PT597110X687430LKGECATA
                event_type:
                  type: string
                  example: PAYMENT.CAPTURE.COMPLETED
                resource:
                  type: object
                  properties:
                    id:
                      type: string
                      example: 1A12345678901234567
                    amount:
                      type: object
                      properties:
                        currency_code:
                          type: string
                          example: USD
                        value:
                          type: string
                          example: 10.00
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Webhook processed successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/payments/paypal/refund:
    post:
      tags:
        - Payments
      summary: Refund PayPal payment
      description: Process a refund for a PayPal payment (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - payment_id
              properties:
                payment_id:
                  type: string
                  format: uuid
                  example: 123e4567-e89b-12d3-a456-426614174000
                amount:
                  type: number
                  format: decimal
                  example: 50.00
                  description: Partial refund amount (optional, full refund if not provided)
                reason:
                  type: string
                  example: Customer requested refund
      responses:
        '200':
          description: Refund processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Refund processed successfully
                  data:
                    type: object
                    properties:
                      refund_id:
                        type: string
                        example: 1B12345678901234567
                      status:
                        type: string
                        example: COMPLETED
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Validation failed
            details:
              - field: email
                message: Invalid email format
    Unauthorized:
      description: Unauthorized - Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Access token required
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Insufficient permissions
    NotFound:
      description: Not found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Resource not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Internal server error
...